+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [0.28.0]
+### Changed
+- Support for the new `toml-rb` release;
+- Updated dev-dependencies;
+
 ## [0.27.0] - 2022-01-12
 ### Changed
 - Drop Ruby 2.5 support.